+++ assets/js/admin/types.js
@@ -163,7 +163,7 @@ export default function types() {
 	function cancelAddCourierNoticeType(event) {
 		event.preventDefault();
 
-		var $target = $('#courier-notice-type-new');
+		let $target = $('#courier-notice-type-new');
 
 		$target.fadeOut('fast').promise().done(function () {
 			$('#add-courier-notice-type').removeAttr('disabled').removeClass('disabled');
@@ -192,7 +192,7 @@ export default function types() {
 		var input = $.extend(true, options, inputTemplate);
 
 		displayNewCourierNoticeTypeTemplate(input);
-		// }
+				// }
 	}
 
 	/**
@@ -338,8 +338,12 @@ export default function types() {
 	 */
 	function displayNewCourierNoticeTypeTemplate(item) {
 
-		var $noticeRow = $( courierNoticeTypeTemplate.map( render(item)).join('') );
-		$('table.courier_notice_page_courier tbody').append( $( $noticeRow ) );
+		let $noticeRow = $( courierNoticeTypeTemplate.map( render(item)).join('') );
+			$noticeRow = $( $noticeRow );
+
+		$('table.courier_notice_page_courier tbody').append( $noticeRow );
+
+		setupTypeEditing( $noticeRow );
 	}
 
 	/**
@@ -372,7 +376,7 @@ export default function types() {
 				$target.fadeOut('fast').promise().done(function () {
 					$('#add-courier-notice-type').removeAttr('disabled').removeClass('disabled');
 
-					for (var fragment in response.fragments) {
+					for ( const fragment in response.fragments ) {
 						$(fragment).html(response.fragments[fragment]);
 					}
 				});
@@ -412,7 +416,7 @@ export default function types() {
 			if ( response && response.fragments ) {
 				$target.fadeOut( 'fast' ).promise().done(
 					function() {
-						for ( let fragment in response.fragments) {
+						for ( let fragment in response.fragments ) {
 							$( fragment ).html( response.fragments[ fragment ] );
 						}
 					}