+++ src/Helper/Functions.php
@@ -26,12 +26,14 @@ defined( 'ABSPATH' ) || exit;
  *
  * @return bool
  */
-function courier_notices_add_notice( $notice = '', $types = array( 'Info' ), $global = false, $dismissible = true, $user_id = 0 ) {
+function courier_notices_add_notice( $notice = '', $types = array( 'Informational' ), $global = false, $dismissible = true, $user_id = 0, $style = 'Informational', $placement = array( 'header' ) ) {
+
+	$user_id = empty( $user_id ) ? get_current_user_id() : intval( $user_id );
 
 	$notice_args = array(
 		'post_type'    => 'courier_notice',
 		'post_status'  => 'publish',
-		'post_author'  => empty( $user_id ) ? get_current_user_id() : intval( $user_id ),
+		'post_author'  => $user_id,
 		'post_name'    => uniqid(),
 		'post_title'   => empty( $notice['post_title'] ) ? '' : sanitize_text_field( $notice['post_title'] ),
 		'post_excerpt' => empty( $notice['post_excerpt'] ) ? '' : wp_kses_post( $notice['post_excerpt'] ),
@@ -48,11 +50,18 @@ function courier_notices_add_notice( $notice = '', $types = array( 'Info' ), $gl
 	}
 
 	if ( $notice_id = wp_insert_post( $notice_args ) ) { // phpcs:ignore
+
 		if ( ! is_array( $types ) ) {
 			$types = explode( ',', $types );
 		}
 
-		wp_set_object_terms( $notice_id, $types, 'courier_type', false );
+		if ( ! is_array( $placement ) ) {
+			$placement = explode( ',', $placement );
+		}
+
+		wp_set_object_terms( $notice_id, $style, 'courier_style', true );
+		wp_set_object_terms( $notice_id, $types, 'courier_type', true );
+		wp_set_object_terms( $notice_id, $placement, 'courier_placement', true );
 
 		if ( $global ) {
 			wp_set_object_terms( $notice_id, array( 'Global' ), 'courier_scope', false );
@@ -63,11 +72,15 @@ function courier_notices_add_notice( $notice = '', $types = array( 'Info' ), $gl
 			wp_cache_delete( 'global-persistent-notices', 'courier-notices' );
 		}
 
+		if ( ! empty( $user_id ) ) {
+			wp_set_object_terms( $notice_id, array( 'User' ), 'courier_scope', false );
+		}
+
 		if ( $dismissible ) {
 			update_post_meta( $notice_id, '_courier_dismissible', 1 );
 		}
 
-		return true;
+		return $notice_id;
 	} else {
 		return false;
 	}
@@ -239,7 +252,6 @@ function courier_notices_display_modals( $args = array() ) {
 
 	$output       = apply_filters( 'courier_notices_modal', $output );
 	$allowed_html = Utils::get_safe_markup();
-
 	echo wp_kses( $output, $allowed_html );
 }
 
@@ -403,4 +415,32 @@ function courier_notices_get_css() {
 	}
 
 	return wp_strip_all_tags( $courier_css );
+}
+
+/**
+ * Display the title for a Courier Notice
+ *
+ * @since 1.3.0
+ *
+ * @param        $title
+ * @param string $before
+ * @param string $after
+ * @param bool   $echo
+ *
+ * @return mixed|string|void
+ */
+function courier_notices_the_notice_title( $title, $before = '', $after = '', $echo = true ) {
+
+	if ( 0 === strlen( $title ) ) {
+		return '';
+	}
+
+	$title = $before . $title . $after;
+	$title = apply_filters( 'courier_notices_the_notice_title', $title );
+
+	if ( $echo ) {
+		echo wp_kses_post( $title );
+	} else {
+		return $title;
+	}
 }