+++ src/Model/Courier_Notice/Data.php
@@ -297,8 +297,6 @@ class Data {
 
 		$post_list = array_merge( $global_posts, $global_dismissible_posts );
 
-		Utils::courier_notices_debug_log( $post_list, 'Query', false );
-
 		// Prioritize Persistent Global Notes to the top by getting them separately and putting them at the front of the line.
 		if ( true === $args['prioritize_persistent_global'] ) {
 			$persistent_global = $this->get_persistent_global_notices(
@@ -344,8 +342,6 @@ class Data {
 		$query_args          = apply_filters( 'courier_notices_display_notices_query', $query_args, $ajax_post_data );
 		$query_args          = wp_parse_args( $args, $query_args );
 		$final_notices_query = new \WP_Query( $query_args );
-
-		Utils::courier_notices_debug_log( $final_notices_query, 'Query', false );
 
 		return ( $final_notices_query->have_posts() ) ? $final_notices_query->posts : array();
 	}
@@ -446,5 +442,68 @@ class Data {
 		$notices_query = new \WP_Query( $query_args );
 
 		return $notices_query->posts;
+	}
+
+	/**
+	 * Get all the relevant meta associated with a notice
+	 *
+	 * @since 1.3.0
+	 *
+	 * @param int $courier_notice_id PostID of the Notice
+	 */
+	public function get_notice_meta( int $courier_notice_id ) {
+
+		$is_dismissible = get_post_meta( $courier_notice_id, '_courier_dismissible', true );
+		$show_title     = get_post_meta( $courier_notice_id, '_courier_show_title', true );
+		$hide_title     = get_post_meta( $courier_notice_id, '_courier_hide_title', true );
+		$courier_style  = get_the_terms( $courier_notice_id, 'courier_style' ); // Get the style associated with the notice
+		$courier_type   = get_the_terms( $courier_notice_id, 'courier_type' );  // Get the type associated with the notice (typically for informational notices)
+		$courier_icon   = get_term_meta( $courier_type[0]->term_id, '_courier_type_icon', true );
+		// Get all the options for showing the title by default
+		$courier_design_options         = get_option( 'courier_design', array() );
+		$global_show_title_rules        = $courier_design_options['enable_title'];
+		$notice_style_global_show_title = in_array( $courier_style[0]->slug, $global_show_title_rules, true );
+
+		// If the notice style is set to show the title by default
+		if ( ! empty( $courier_style ) ) {
+			// Force the title to show if the type supports it.
+			if ( $notice_style_global_show_title ) {
+				$show_hide_title = 'show';
+			}
+		} else {
+			if ( ! empty( $show_title ) ) {
+				$show_hide_title = 'show';
+			}
+
+			// If we are not forcing show/hide of the title on the notice itself, check to see if we have a default option set.
+			if ( empty( $show_title ) && empty( $hide_title ) ) {
+				$show_hide_title = 'hide';
+			}
+		}
+
+		// Override the show global toggle and force this notice to hide.
+		if ( ! empty( $hide_title ) ) {
+			$show_hide_title = 'hide';
+		}
+
+		if ( ! empty( $show_title ) ) {
+			$show_hide_title = 'show';
+		}
+
+		// Failsafe to hide the title
+		if ( empty( $show_hide_title ) ) {
+			$show_hide_title = 'hide';
+		}
+
+		$notices_meta = array(
+			'is_dismissible'  => ( $is_dismissible ) ? $is_dismissible : false,
+			'show_hide_title' => ( $show_hide_title ) ? $show_hide_title : 'hide',
+			'style'           => $courier_style,
+			'type'            => $courier_type,
+			'icon'            => $courier_icon,
+			'is_confirmation' => has_term( 'gform-confirmation', 'courier_visibility_rules', $courier_notice_id ),
+		);
+
+		return apply_filters( 'courier_notices_notice_meta', $notices_meta );
 	}
 }