+++ templates/notice-popup-modal.php
@@ -3,11 +3,12 @@
 	$icon         = ( isset( $icon ) && '' !== $icon ) ? 'icon-' . $icon : '';
 	$dismissible  = ( ! empty( $dismissible ) ) ? 'data-closable' : '';
 ?>
-<div class="courier-modal-overlay" style="display:none;">
-	<div class="courier-notices modal <?php echo esc_attr( $notice_class ); ?>" data-courier-notice-id="<?php echo esc_attr( $notice_id ); ?>" <?php esc_attr( $dismissible ); ?>>
-		<?php if ( $dismissible ) : ?>
-			<a href="#" class="courier-close close">&times;</a>
-		<?php endif; ?>
-		<div class="courier-content"><?php echo wp_kses_post( $notice_content ); ?></div>
-	</div>
+<div class="courier-notices modal <?php echo esc_attr( $notice_class ); ?>" <?php esc_attr( $dismissible ); ?> data-courier-notice-id="<?php echo esc_attr( $notice_id ); ?>">
+	<?php if ( $dismissible ) : ?>
+		<a href="#" class="courier-close modal-close close">&times;</a>
+	<?php endif; ?>
+	<?php if ( 'hide' !== $show_hide_title ) : ?>
+		<h6><?php echo wp_kses_post( $notice_title ); ?></h6>
+	<?php endif; ?>
+	<div class="courier-content"><?php echo wp_kses_post( $notice_content ); ?></div>
 </div>